Modify SQL scripts from MySQL to PostgreSQL

+++ src/main/resources/db/migration/V20201130212152__create_offer_table.sql
@@ -1,9 +1,9 @@
 CREATE TABLE IF NOT EXISTS offer (
        id BIGINT PRIMARY KEY NOT NULL,
        request_id BIGINT NOT NULL,
-       total_value DOUBLE,
+       total_value DOUBLE PRECISION,
        number_of_installments INTEGER,
-       installment_value Double,
+       installment_value DOUBLE PRECISION,
 
        created_at TIMESTAMP NOT NULL DEFAULT CURRENT_TIMESTAMP,
        updated_at TIMESTAMP,
